move edit from web test from permissions to database test
